tests: Move crate yanking tests to `krate::yanking` module
