+++ src/tests/http-data/krate_yanking_yank_works_as_intended
